Copter: reorganise the iris launch folder
- Move iris specific launch declaration into robot/iris.launch.
- Rename the world model and launch files.
- Add placeholder for models (empty folder).
- Zoom to model in rviz config.
- Update README
