+++ ardupilot_gz_description/models/README.md
@@ -1,0 +1,1 @@
+# Placeholder for `ardupilot_gz_description` models.
